requirements.txt
# Face Recognition Door Unlock System
# Core dependencies

opencv-python>=4.8.0
numpy>=1.24.0
face-recognition>=1.3.0
dlib>=19.24.0
scipy>=1.11.0
Pillow>=10.0.0

# For anti-replay with image hashing
imagehash>=4.3.0
